Avoid returning types containing ? from glb/lub
when an otherwise equivalent type without wildcards is available.
This improves unification of type variables in the following way:

It's possible for a type var `?A` to be constrained by a "partial"
type containing a wildcard such as `(Int, ?)`. Since for any type
`T` both `T <: ?` and `? <: T`, the "partial" type could shadow a
more informative type such as `(Int, String)` in `glb` and `lub`.
The net result being that the success or failure of unification
depended on the order of constraints in `?A`.

This commit incorporates the simple heuristic of preferring types
that don't contain wildcards in `glb` and `lub`.

Fixes scala/bug#10686