Defer some of parameter alias computation until after typer
There is a data race between Typer::computeParamAliases running for
the params in class b and c. If c is typechecked first, it isn't
able to know that b.x is aliased by a.x, and as leaves c.x
unaliased (ie, we spend a field on it).

The solution is to defer some of the work of
Typer::computeParamAliases, so that during typechecking we just do the
tree inspection, but defer the part that requires knowledges of the
parent constructor until a later phase (e.g. refchecks or superaccessors).
That might be best structured in an info transform (even if it is just
mutating Symbol.referenced)