Be less intrusive with customization of baseDirectory
In our attempt to make forked project take on the working directory
of the root project, rather than of their module, we ended up
causing a clash in the directory used by the incremental compiler
to backup classfiles, which uses a value relative to that base.

The symptom was an intermitten error such as:

```
[error] ## Exception when compiling 0 sources to /Users/jz/code/scala-sbt-1/target/partest-extras/test-classes
[error] Could not create directory /Users/jz/code/scala-sbt-1/target/scala-2.13.0-M3/classes.bak
```

This commit adjusts ForkOptions instead to more directly
acheive our goal.