Avoid as-seen-from from treating eta-expanded type vars as class tparams
Since #6069, it is possible to have type variables compared to the eta
expansion of a class type, which is `[CLASSTPARAM] => Class[CLASSTPARAM]`.

After separate compilation, the owner this higher kinded type param
is adjusted to the root class (see the comments and special cases in
Pickler.localizedOwner), which makes it show up to certain subsequent
as-seen-from maps as a type parameter of an enclosing class. SBT's
ExtractAPI phase runs into this problem.

This commit pins down the problem with a test and offers a straw-man
fix, which clones the type params at a different owner (currently,
NoOwner, but with a bit more refactoring this should be context.owner)
when type when instantiating the type variables.

An alternative might be to clone during eta-expansion (although to
which owner? NoSymbol? A user provided one? What does this mean for
our strategy of caching normalized types for performance?)

A third fix would be to do the `cloneInfo` before registering the
bound in `TypeComparers`. We'd need to refactor somewhat to thread
through `context.owner` to that code, however.

Fixes scala/bug#t10762