Eta-expanded tycon's type params need own owner
Since #6069, it is possible to have type variables compared to the eta
expansion of a class type, which is `[CLASSTPARAM] => Class[CLASSTPARAM]`.

After separate compilation, the owner of this type constructor param
is adjusted to the root class (see the comments and special cases in
Pickler.localizedOwner), which makes it show up to certain subsequent
as-seen-from maps as a type parameter of an enclosing class. SBT's
ExtractAPI phase runs into this problem.

Eta expansion of the tycon type ref uses an owner (NoSymbol) that
will not become a class after pickling.

Fixes scala/bug#10762