Don't evaluate devWarning conditions unless under -Xdev
The shape of `devWarningIf` is such that the inliner produces bytecode
similar to

```scala
if (global.isDeveloper) {
  if (condition) {
    devWarning(() => msg)
  }
}
```

so that the closure elimination and the condition evaluation are both
guarded by the `isDeveloper` check. (This does mean that the bytecode
for the condition gets emitted into the containing method.)

Several of these `devWarning` conditions were based on list traversals
or other non-constant-time conditions, so this should speed that up
a bit.