Typers: by-pass the unzipped list of pairs.
The modified code in the Typers file was creating a list of pairs,
by using a `map2` function, which was immediately _unzipped_.
Thus, the first list was a linear number of unneeded allocations.

To avoid that intermediate list, we change the algorithm to directly
generate the two lists on a single pass. To do this, we use mutable
ListBuffer, and  replace the `map2` with a `foreach`.