s.c.i.HashMap (HashSet) concat and removedAll implementations use shallow mutations
during concatenation with a non-hashmap (hashset) IterableOnce,
the hashmap (hashset) will perform shallow mutations on any newly
created nodes. Nodes are known to be newly created if:

* they are the root node, and the root node has been returned from
updated/removed, and that root node is not `eq` to the previous
root node

* additionally, during concat, any newly created nodes which are
immediate descenants are tracked via a new bitmap `shallowlyMutableNodeMap`,
which tracks any newly-created nodes in the level beneath the root,
so that later in the concatenation (for later elements), those
nodes can also be mutated rather than immutably updated.

Revert functionality changes from #7718