Improvements to async phase
Passes the correct owner symbol to `useFields` to fix a bug
with local modules.

Refactors to store state about the current async block in a var
which allows us to mix in all the implementation once into
AsyncPhase.

Remove some testing FutureSystems. We'll refactor the tests
to use ScalaConcurrentFutureSystem.

Allow a custom front end to supply an alternative FutureSystem
attached to the block defining the state machine wrapper class.
Demonstrate this with a test case for an annotation driven
front end that targets an alternative Future implementation.

Remove some now-unused customization points from FutureSystem.
For example, we don't need to abstract over the choice of
base class for the state machine as a custom front end now
creates the state machine tree directly.

FutureSystem now deals directly in Trees, rather than in
`c.Expr`. I've kept a light weight option (a type alias
`type Expr[T] = Tree`) so we can have a little compiler support
for getting the types right and documentation for
`FutureSystem` implementors.

Refactor the ANF transform (and others) to directly be a
TypingTransformer.

Remove the need to subclass Function0 in the state machine
by kicking things off with `Future.unit.onComplete(stateMachine)`.
This reduces bloat in the generated code due to Function0
specialization.

I've also elminated the macro application in the early transform
in favour of wrapping the argument in `Block(<arg> :: Nil, q"()"}`.
This ran into some "pure expression has no effect" warnings which
I've suppressed with a new tree annotation.

Most of the `auto_` test cases are converted to use the
standard scala.concurrent version of async. Now that the
transform is always happening post-erasure, there is no need to
keep those tests under the annotation driven async test (which
used to be the only way to get post-typer expansion).

A few test can only be expressed with the annotation driven
version of async, namely those that have an async boundary
in an extractor. These have been moved into the
`AnnotationDrivenAsyncTest`.