Reduce number of states in async state machine
We used to lift the control flow for any pattern match or
if/else into fine grained states if it enclosed even a single
await call.

Instead, we now keep as much of the control flow as possible
in the current state. The program point after the control flow
statement is lifted into a new state iff it is reachable from
any branch that contains an await.

LabelDef-s representing cases and the matchEnd of patterns are now
only lifted into states if they are the target of a label jump
from a resumption handler.

To make this analysis possible, we form a context chain of
AsyncBlockBuilders to let the translation of the label jumps compare
the current state to that current state of the enclosing block builder
that holds the LabelDefs. If these differ, we can conclude that we
have crossed an async boundary and mark the LabelDef for lifting
into state.

Here's how the simplification helps:

https://gist.github.com/retronym/50839a051db9b6d1fcd0719d99ecbfc2

Refactorings along the way:

  - introduce a ThicketTransformer to transform `Tree => List[Tree]`
  - make `symLookup` more convenient to access
  - add some internal flags to disable state compaction and field
    nulling which help to debug the phase
  - unify onto a single `AsyncState` class. Different use cases
    are served by eagerly adding additional logic, rather than
    having this added later on in the variants of
    `mkCompletionHandler`
  - detect and remove empty states in `compactStates`. It seems easier
    now to do this than to avoid constructing them.
  - Try a bit harder to avoid redundant `()` in stats.
  - Move synthesis of async block generation into
    StateBuilder.build (formrly known as resultSimple). This is
    now the only factory method on StateBuilder.
  - Fix propagation of the enclosing subsequent state into
    nested expression positions
  - In combination, this lets us perform async block completion
    inline in all terminal states, there is no need to write the
    result to lifted field and make a state transition.
  - Remove the `currTree` var in favour of delegating to
    the field in the `stateBuilder` var
  - Encapsulate state transition bookkeeping