Reinstate live variables test and fix null assignment
While looking more closely at the implementation, I
realised our approach of nulling dead variables at the
start of the subsequent state was wrong, we retained
the references for longer than needed.

I've moved the nulling into the conclusion of the
state of last usages.

Remove the warnings test which isn't relevant with post-typer ANF.