Remove AST builders from FutureSystem extension point
Now that we require custom integrations of async to
create their own state machine wrapper in an early phase,
require that the state machine implements a set of
methods to deal with the Try and Future types being
used.

As shown in the test `AnnotationDrivenAsync`, these
can be inherited from a base class.