Fix case-apply rewrite in case of private constr and overloaded apply
In the example, `apply` is overloaded.

Overloading resolution uses `setSymbol` and `setType` once it has
selected a winner.

https://github.com/scala/scala/blob/v2.13.14/src/compiler/scala/tools/nsc/typechecker/Infer.scala#L1530

In the case of a `TypeApply` tree:
  - `setSymbol` delegates to the nested `Apply` tree
  - `setType` sets the type of the `TypeApply`

This means that the nested `Apply` tree keeps its OverloadedType.
This tripped up `fun.tpe.finalResultType.typeSymbol` in RefChecks (NoSymbol).

It's not obvious how to assign the right type to the `Apply` tree, so
I left it alone.